:arrow_forward: Authentication to routes
Authentication was added to the routes to redirect them in case the user is or is not logged in, because only logged in users can see the routes of `/notes` and those who are not, can see the routes of `/users`.

+++ src/controllers/notes.controllers.js
@@ -8,18 +8,23 @@ notesCtrl.renderNoteForm = (req, res) => {
 notesCtrl.createNewNote = async (req, res) => {
   const { title, description } = req.body
   const newNote = new Note({ title, description })
+  newNote.user= req.user.id
   await newNote.save()
   req.flash('success_msg', 'Note added successfully')
   res.redirect('/notes')
 }
 
 notesCtrl.renderNotes = async (req, res) => {
-  const notes = await Note.find()
+  const notes = await Note.find({user: req.user.id}).sort({createdAt: 'desc'})
   res.render('notes/all-notes', { notes })
 }
 
 notesCtrl.renderEditForm = async (req, res) => {
   const note = await Note.findById(req.params.id)
+  if(req.user.id !== note.user){
+    req.flash('error_msg', 'Not authorized')
+    return res.redirect('/notes')
+  }
   res.render('notes/edit-notes', { note })
 }
 
@@ -31,6 +36,11 @@ notesCtrl.updateNote = async (req, res) => {
 }
 
 notesCtrl.deleteNote = async (req, res) => {
+  const note = await Note.findById(req.params.id)
+  if(req.user.id !== note.user){
+    req.flash('error_msg', 'Not authorized')
+    return res.redirect('/notes')
+  }
   await Note.findByIdAndDelete(req.params.id)
   req.flash('success_msg', 'Note deleted successfully')
   res.redirect('/notes')
